% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergastR-core.R
\name{constructorStandings.df}
\alias{constructorStandings.df}
\title{Get dataframe for constructor final standings by year or race}
\usage{
constructorStandings.df(year, race = "")
}
\arguments{
\item{character}{constructorRef}
}
\value{
dataframe containing standings for each driver at the end of each year
}
\description{
\code{constructorStandings.df}
}
