% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergastR-core.R
\name{driverCodeExists}
\alias{driverCodeExists}
\title{Test whether a driver code exists and if so, return it}
\usage{
driverCodeExists(x)
}
\arguments{
\item{list}{item that may or may not contain a code sublist}
}
\value{
driver code
}
\description{
\code{driverCodeExists}
}
