% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergastR-core.R
\name{driverResults.df}
\alias{driverResults.df}
\title{Get dataframe containing results by driver}
\usage{
driverResults.df(year, driverRef = NA)
}
\arguments{
\item{integer}{season year}

\item{character}{driverRef reference code for specified driver}
}
\value{
dataframe containing results data for a particular driver in a particular year
}
\description{
\code{driverResults.df}
}
