% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergastR-core.R
\name{driversData.df}
\alias{driversData.df}
\title{Get dataframe for drivers by year}
\usage{
driversData.df(year)
}
\arguments{
\item{integer}{season year for required data}
}
\value{
dataframe containing race data for each year
}
\description{
\code{driversData.df}
}
