% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergastR-core.R
\name{ergastR}
\alias{ergastR}
\alias{getNum}
\title{Wrapper for ergast.com JSON API}
\usage{
getNum(x)
}
\arguments{
\item{object}{item to be converted to a numeric}
}
\value{
string or factor cast as numeric
}
\description{
Wrap calls to \code{ergast.com} JSON API and return R dataframes.
}
\author{
Tony Hirst (@psychemedia)
Convert a string to a numeric

\code{getNum}
}
