% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergastR-core.R
\name{formatLapData}
\alias{formatLapData}
\title{Format laps data}
\usage{
formatLapData(rd)
}
\arguments{
\item{object}{containing laps data}
}
\value{
dataframe containing laps data
}
\description{
\code{formatLapData}
}
