% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergastR-core.R
\name{formatPitsData}
\alias{formatPitsData}
\title{Extract and format pits data from JSON object}
\usage{
formatPitsData(rd.pits)
}
\arguments{
\item{object}{containg race data}
}
\value{
object containing laps data
}
\description{
\code{formatPitsData}
}
