% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergastR-core.R
\name{getDriverResultsByYear.URL}
\alias{getDriverResultsByYear.URL}
\title{Get URL for results by year and driver}
\usage{
getDriverResultsByYear.URL(year, driverRef = NA, format = "json")
}
\arguments{
\item{integer}{season year}

\item{character}{driverRef}

\item{character}{data format (json, XML)}
}
\value{
a URL
}
\description{
*TO DO - this function is hacky in that an explicit maximum limit value is et rather than using a proper paged results handler*
\code{getDriverResultsByYear.URL}
}
