% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergastR-core.R
\name{getDriversByYear.URL}
\alias{getDriversByYear.URL}
\title{Get URL for results by race-and-year}
\usage{
getDriversByYear.URL(year, format = "json")
}
\arguments{
\item{integer}{season year}

\item{character}{data format (json, XML)}
}
\value{
a URL
}
\description{
\code{getDriversByYear.URL}
}
