% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergastR-core.R
\name{getJSONbyURL}
\alias{getJSONbyURL}
\title{Get JSON data}
\usage{
getJSONbyURL(URL)
}
\arguments{
\item{character}{URL for data request}
}
\value{
JSON data from ergast API
}
\description{
*TO DO - really should have a handler that supports cacheing*
\code{getJSONbyURL}
}
