% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergastR-core.R
\name{getLapsByYearRaceDriver.URL}
\alias{getLapsByYearRaceDriver.URL}
\title{Get URL for laps by race-and-year-and-driver}
\usage{
getLapsByYearRaceDriver.URL(year, raceNum, driverId, format = "json")
}
\arguments{
\item{integer}{season year}

\item{integer}{race number in year}

\item{character}{driver}

\item{character}{data format (json, XML)}
}
\value{
a URL
}
\description{
*TO DO - this function is hacky in that an explicit maximum limit value is et rather than using a proper paged results handler*
\code{getLapsByYearRaceDriver.URL}
}
