% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergastR-core.R
\name{getLapsData.path}
\alias{getLapsData.path}
\title{Extract Laps data from race data JSON object}
\usage{
getLapsData.path(rd.laps)
}
\arguments{
\item{object}{containg race data}
}
\value{
object containing laps data
}
\description{
\code{getLapsData.path}
}
