% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergastR-core.R
\name{getPitsByYearRace.URL}
\alias{getPitsByYearRace.URL}
\title{Get URL for pits by race-and-year}
\usage{
getPitsByYearRace.URL(year, raceNum, format = "json")
}
\arguments{
\item{integer}{season year}

\item{integer}{race number in year}

\item{character}{data format (json, XML)}
}
\value{
a URL
}
\description{
\code{getPitsByYearRace.URL}
}
