% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergastR-core.R
\name{getRaceResultsByYearRace.URL}
\alias{getRaceResultsByYearRace.URL}
\title{Get URL for results by race-and-year}
\usage{
getRaceResultsByYearRace.URL(year, raceNum, format = "json")
}
\arguments{
\item{integer}{season year}

\item{integer}{race number in year}

\item{character}{data format (json, XML)}
}
\value{
a URL
}
\description{
*TO DO - this function is hacky in that an explicit maximum limit value is et rather than using a proper paged results handler*
\code{getRaceResultsByYearRace.URL}
}
