% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergastR-core.R
\name{getRacesDataByYear.URL}
\alias{getRacesDataByYear.URL}
\title{Get URL for races by year}
\usage{
getRacesDataByYear.URL(year, format = "json")
}
\arguments{
\item{integer}{season year for required data}

\item{character}{data format (json, XML)}
}
\value{
a URL
}
\description{
\code{getRacesDataByYear.URL}
}
