% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergastR-core.R
\name{lapsData.df}
\alias{lapsData.df}
\title{Generate dataframe containing lap data for a given race}
\usage{
lapsData.df(year, raceNum, format = "json", con = NA)
}
\arguments{
\item{integer}{season year for required data}

\item{integer}{round number for required data}

\item{character}{data format (json, XML)}
}
\value{
dataframe containing lap data for a specific race
}
\description{
\code{lapsData.df}
}
