% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergastR-core.R
\name{lapsDataDriver.df}
\alias{lapsDataDriver.df}
\title{Generate dataframe containing lap data for a specified driver}
\usage{
lapsDataDriver.df(year, raceNum, driver, format = "json")
}
\arguments{
\item{integer}{season year for required data}

\item{integer}{round number for required data}

\item{character}{driverId for specified driver}

\item{character}{data format (json, XML)}
}
\value{
dataframe containing lap data for a specific race
}
\description{
\code{lapsDataDriver.df}
}
