% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergastR-core.R
\name{pitsData.df}
\alias{pitsData.df}
\title{Generate dataframe containing pit data for a specified year and round}
\usage{
pitsData.df(year, raceNum, format = "json")
}
\arguments{
\item{integer}{season year for required data}

\item{integer}{round number for required data}

\item{character}{data format (json, XML)}
}
\value{
dataframe containing lap data for a specific race
}
\description{
\code{pitsData.df}
}
