% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergastR-core.R
\name{qualiResults.df}
\alias{qualiResults.df}
\title{Parse dataframe containing qualifying results}
\usage{
qualiResults.df(year = NA, raceNum = NA, driverRef = NA,
  constructorRef = NA, format = "json", con = NA)
}
\arguments{
\item{integer}{season year}

\item{integer}{round number for required data}

\item{character}{driverRef reference code for specified driver}

\item{character}{constructor reference code for specified constructor}

\item{character}{data format (json, XML)}

\item{character}{connection string for database; if this is not present use ergast API}
}
\value{
dataframe containing qualifying results
}
\description{
\code{qualiResults.df}
}
