% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergastR-core.R
\name{qualiResultsParse.df}
\alias{qualiResultsParse.df}
\title{Parse dataframe containing qualifying results}
\usage{
qualiResultsParse.df(url)
}
\arguments{
\item{character}{url of the API call}
}
\value{
dataframe containing qualifying results
}
\description{
\code{qualiResultsParse.df}
}
