% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergastR-core.R
\name{racesData.df}
\alias{racesData.df}
\title{Get dataframe for races by year}
\usage{
racesData.df(year)
}
\arguments{
\item{integer}{season year for required data}
}
\value{
dataframe containing race data for each year
}
\description{
\code{racesData.df}
}
