% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergastR-core.R
\name{raceWinner}
\alias{raceWinner}
\title{Get race winner by season and year}
\usage{
raceWinner(year, raceNum)
}
\arguments{
\item{integer}{season year}

\item{integer}{race number}
}
\value{
driverId for winner of a particular race
}
\description{
\code{raceWinner}
}
