% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergastR-core.R
\name{resultsData.df}
\alias{resultsData.df}
\title{Get dataframe for races by year}
\usage{
resultsData.df(year, raceNum)
}
\arguments{
\item{integer}{season year}

\item{integer}{race number in season}
}
\value{
dataframe containing results data for a particular race
}
\description{
\code{resultsData.df}
}
