% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergastR-core.R
\name{seasonStandings}
\alias{seasonStandings}
\title{Get dataframe for season standings by year}
\usage{
seasonStandings(year, race = "")
}
\arguments{
\item{integer}{season year}

\item{integer}{race number}
}
\value{
dataframe containing standings for each year, by race
}
\description{
\code{seasonStandings.df}
}
