% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergastR-core.R
\name{timeInS}
\alias{timeInS}
\title{Generate time in seconds from min:seconds format}
\usage{
timeInS(tStr)
}
\arguments{
\item{string}{time in format minutes:seconds}
}
\value{
time in seconds as numeric
}
\description{
\code{timeInS}
}
